pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-policy-testkit"
version = "0.1.0"
description = "CI-first policy and security testing for MCP servers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "jansona" }
]
dependencies = [
  "jinja2>=3.1",
  "pydantic>=2.7",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5.0",
]

[project.scripts]
mcp-policy-testkit = "mcp_policy_testkit.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["UP006", "UP035", "UP042", "UP045"]
